requirements.txt
aws-cdk-lib==2.26.0
constructs>=10.0.0,<11.0.0
boto3
amazon-textract-idp-cdk-constructs>=0.0.4
schadem-tidp-manifest==0.0.8
